requirements.txt
customtkinter
tk
pymongo
openai
requests
python-dotenv
bcrypt
ttkthemes
packaging
threading